Move NameManager and related classes to own package
